refactor: restructure project paths, update Next.js config, and adjust build output
