+++ lib/getToken.ts
