+++ next.config.ts
@@ -1,0 +1,12 @@
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+  reactStrictMode: true,
+  output: "export",
+  images: {
+    unoptimized: true,
+  },
+  distDir: "dist",
+};
+
+export default nextConfig;
